# postify-dev/api_spec.yaml
openapi: 3.0.3
info:
  title: Postify API
  description: Producer/Consumer documentation for the Postify platform.
  termsOfService: https://apiinfo.madface.info/tos.html
  contact:
    name: Postify API Support
    url: https://apiinfo.madface.info/support.html
  license:
    name: CC BY-NC 4.0
    url: https://creativecommons.org/licenses/by-nc/4.0
  version: 0.3.0
servers:
  - url: https://postify.charliewei.com/4537/termproject/api/v1/
    description: General purpose
paths:
...
  "/project/solution":
    get:
      summary: gets a project solution
      responses:
        "200":
          description: Solution
          content:
            application/json:
              schema:
                title: "Solution"
                type: object
                properties:
                  solutionID:
                    type: integer
                  projID:
                    type: integer
                  userID:
                    type: integer
                  solDesc:
                    type: string
                  solGitLink:
                    type: string
                  solFileURL:
                    type: string
                  solUpVote:
                    type: integer
                  solDownVote:
                    type: integer
                  solCreateDate:
                    type: string
                  solEditDate:
                    type: string
                  
    put:
      summary: Updates a project solution
      requestBody:
        description: Updated olution details
        required: true
        content:
          application/json:
            schema:
              title: "Solution"
              type: object
              properties:
                solDesc:
                  type: string
                solGitLink:
                  type: string
                solFileURL:
                  type: string
                solUpVote:
                  type: string
                solDownVote:
                  type: string
                solutionID:
                  type: string
      responses:
        "200":
          description: Deleted solution
        "404":
          description: Unable to delete solution        